+++ sdk/mediaservices/arm-mediaservices/src/azureMediaServices.ts
@@ -7,15 +7,9 @@
  */
 
 import * as coreClient from "@azure/core-client";
-import * as coreRestPipeline from "@azure/core-rest-pipeline";
 import * as coreAuth from "@azure/core-auth";
 import {
   AccountFiltersImpl,
-  OperationsImpl,
-  MediaservicesImpl,
-  PrivateLinkResourcesImpl,
-  PrivateEndpointConnectionsImpl,
-  LocationsImpl,
   AssetsImpl,
   AssetFiltersImpl,
   TracksImpl,
@@ -32,11 +26,6 @@ import {
 } from "./operations";
 import {
   AccountFilters,
-  Operations,
-  Mediaservices,
-  PrivateLinkResources,
-  PrivateEndpointConnections,
-  Locations,
   Assets,
   AssetFilters,
   Tracks,
@@ -56,6 +45,7 @@ import { AzureMediaServicesOptionalParams } from "./models";
 export class AzureMediaServices extends coreClient.ServiceClient {
   $host: string;
   subscriptionId: string;
+  apiVersion: string;
 
   /**
    * Initializes a new instance of the AzureMediaServices class.
@@ -84,7 +74,7 @@ export class AzureMediaServices extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-mediaservices/11.0.1`;
+    const packageDetails = `azsdk-js-arm-mediaservices/12.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -99,44 +89,16 @@ export class AzureMediaServices extends coreClient.ServiceClient {
       userAgentOptions: {
         userAgentPrefix
       },
-      baseUri:
-        options.endpoint ?? options.baseUri ?? "https://management.azure.com"
+      baseUri: options.endpoint || "https://management.azure.com"
     };
     super(optionsWithDefaults);
-
-    if (options?.pipeline && options.pipeline.getOrderedPolicies().length > 0) {
-      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] = options.pipeline.getOrderedPolicies();
-      const bearerTokenAuthenticationPolicyFound = pipelinePolicies.some(
-        (pipelinePolicy) =>
-          pipelinePolicy.name ===
-          coreRestPipeline.bearerTokenAuthenticationPolicyName
-      );
-      if (!bearerTokenAuthenticationPolicyFound) {
-        this.pipeline.removePolicy({
-          name: coreRestPipeline.bearerTokenAuthenticationPolicyName
-        });
-        this.pipeline.addPolicy(
-          coreRestPipeline.bearerTokenAuthenticationPolicy({
-            scopes: `${optionsWithDefaults.baseUri}/.default`,
-            challengeCallbacks: {
-              authorizeRequestOnChallenge:
-                coreClient.authorizeRequestOnClaimChallenge
-            }
-          })
-        );
-      }
-    }
     // Parameter assignments
     this.subscriptionId = subscriptionId;
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
+    this.apiVersion = options.apiVersion || "2021-12-01";
     this.accountFilters = new AccountFiltersImpl(this);
-    this.operations = new OperationsImpl(this);
-    this.mediaservices = new MediaservicesImpl(this);
-    this.privateLinkResources = new PrivateLinkResourcesImpl(this);
-    this.privateEndpointConnections = new PrivateEndpointConnectionsImpl(this);
-    this.locations = new LocationsImpl(this);
     this.assets = new AssetsImpl(this);
     this.assetFilters = new AssetFiltersImpl(this);
     this.tracks = new TracksImpl(this);
@@ -153,11 +115,6 @@ export class AzureMediaServices extends coreClient.ServiceClient {
   }
 
   accountFilters: AccountFilters;
-  operations: Operations;
-  mediaservices: Mediaservices;
-  privateLinkResources: PrivateLinkResources;
-  privateEndpointConnections: PrivateEndpointConnections;
-  locations: Locations;
   assets: Assets;
   assetFilters: AssetFilters;
   tracks: Tracks;