+++ sdk/mediaservices/arm-mediaservices/src/models/parameters.ts
@@ -13,12 +13,6 @@ import {
 } from "@azure/core-client";
 import {
   AccountFilter as AccountFilterMapper,
-  MediaService as MediaServiceMapper,
-  MediaServiceUpdate as MediaServiceUpdateMapper,
-  SyncStorageKeysInput as SyncStorageKeysInputMapper,
-  ListEdgePoliciesInput as ListEdgePoliciesInputMapper,
-  PrivateEndpointConnection as PrivateEndpointConnectionMapper,
-  CheckNameAvailabilityInput as CheckNameAvailabilityInputMapper,
   Asset as AssetMapper,
   ListContainerSasInput as ListContainerSasInputMapper,
   AssetFilter as AssetFilterMapper,
@@ -95,7 +89,7 @@ export const accountName: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2021-11-01",
+    defaultValue: "2021-12-01",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -144,70 +138,6 @@ export const nextLink: OperationURLParameter = {
   skipEncoding: true
 };
 
-export const apiVersion1: OperationQueryParameter = {
-  parameterPath: "apiVersion",
-  mapper: {
-    defaultValue: "2021-06-01",
-    isConstant: true,
-    serializedName: "api-version",
-    type: {
-      name: "String"
-    }
-  }
-};
-
-export const parameters1: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: MediaServiceMapper
-};
-
-export const parameters2: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: MediaServiceUpdateMapper
-};
-
-export const parameters3: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: SyncStorageKeysInputMapper
-};
-
-export const parameters4: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: ListEdgePoliciesInputMapper
-};
-
-export const name: OperationURLParameter = {
-  parameterPath: "name",
-  mapper: {
-    serializedName: "name",
-    required: true,
-    type: {
-      name: "String"
-    }
-  }
-};
-
-export const parameters5: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: PrivateEndpointConnectionMapper
-};
-
-export const parameters6: OperationParameter = {
-  parameterPath: "parameters",
-  mapper: CheckNameAvailabilityInputMapper
-};
-
-export const locationName: OperationURLParameter = {
-  parameterPath: "locationName",
-  mapper: {
-    serializedName: "locationName",
-    required: true,
-    type: {
-      name: "String"
-    }
-  }
-};
-
 export const filter: OperationQueryParameter = {
   parameterPath: ["options", "filter"],
   mapper: {
@@ -249,17 +179,17 @@ export const assetName: OperationURLParameter = {
   }
 };
 
-export const parameters7: OperationParameter = {
+export const parameters1: OperationParameter = {
   parameterPath: "parameters",
   mapper: AssetMapper
 };
 
-export const parameters8: OperationParameter = {
+export const parameters2: OperationParameter = {
   parameterPath: "parameters",
   mapper: ListContainerSasInputMapper
 };
 
-export const parameters9: OperationParameter = {
+export const parameters3: OperationParameter = {
   parameterPath: "parameters",
   mapper: AssetFilterMapper
 };
@@ -275,7 +205,7 @@ export const trackName: OperationURLParameter = {
   }
 };
 
-export const parameters10: OperationParameter = {
+export const parameters4: OperationParameter = {
   parameterPath: "parameters",
   mapper: AssetTrackMapper
 };
@@ -302,7 +232,7 @@ export const contentKeyPolicyName: OperationURLParameter = {
   }
 };
 
-export const parameters11: OperationParameter = {
+export const parameters5: OperationParameter = {
   parameterPath: "parameters",
   mapper: ContentKeyPolicyMapper
 };
@@ -318,7 +248,7 @@ export const transformName: OperationURLParameter = {
   }
 };
 
-export const parameters12: OperationParameter = {
+export const parameters6: OperationParameter = {
   parameterPath: "parameters",
   mapper: TransformMapper
 };
@@ -334,7 +264,7 @@ export const jobName: OperationURLParameter = {
   }
 };
 
-export const parameters13: OperationParameter = {
+export const parameters7: OperationParameter = {
   parameterPath: "parameters",
   mapper: JobMapper
 };
@@ -350,7 +280,7 @@ export const streamingPolicyName: OperationURLParameter = {
   }
 };
 
-export const parameters14: OperationParameter = {
+export const parameters8: OperationParameter = {
   parameterPath: "parameters",
   mapper: StreamingPolicyMapper
 };
@@ -366,7 +296,7 @@ export const streamingLocatorName: OperationURLParameter = {
   }
 };
 
-export const parameters15: OperationParameter = {
+export const parameters9: OperationParameter = {
   parameterPath: "parameters",
   mapper: StreamingLocatorMapper
 };
@@ -387,7 +317,7 @@ export const liveEventName: OperationURLParameter = {
   }
 };
 
-export const parameters16: OperationParameter = {
+export const parameters10: OperationParameter = {
   parameterPath: "parameters",
   mapper: LiveEventMapper
 };
@@ -402,7 +332,7 @@ export const autoStart: OperationQueryParameter = {
   }
 };
 
-export const parameters17: OperationParameter = {
+export const parameters11: OperationParameter = {
   parameterPath: "parameters",
   mapper: LiveEventActionInputMapper
 };
@@ -423,7 +353,7 @@ export const liveOutputName: OperationURLParameter = {
   }
 };
 
-export const parameters18: OperationParameter = {
+export const parameters12: OperationParameter = {
   parameterPath: "parameters",
   mapper: LiveOutputMapper
 };
@@ -444,12 +374,12 @@ export const streamingEndpointName: OperationURLParameter = {
   }
 };
 
-export const parameters19: OperationParameter = {
+export const parameters13: OperationParameter = {
   parameterPath: "parameters",
   mapper: StreamingEndpointMapper
 };
 
-export const parameters20: OperationParameter = {
+export const parameters14: OperationParameter = {
   parameterPath: "parameters",
   mapper: StreamingEntityScaleUnitMapper
 };