+++ sdk/mediaservices/arm-mediaservices/src/operations/index.ts
@@ -7,11 +7,6 @@
  */
 
 export * from "./accountFilters";
-export * from "./operations";
-export * from "./mediaservices";
-export * from "./privateLinkResources";
-export * from "./privateEndpointConnections";
-export * from "./locations";
 export * from "./assets";
 export * from "./assetFilters";
 export * from "./tracks";